(refactor) Services to a new package services.api
