+++ src/main/java/org/storkforge/barkr/web/controller/ApiAccountController.java
@@ -5,13 +5,9 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import org.storkforge.barkr.domain.AccountService;
-import org.storkforge.barkr.domain.entity.Account;
+import org.storkforge.barkr.domain.services.api.AccountService;
 import org.storkforge.barkr.dto.accountDto.ResponseAccount;
 import org.storkforge.barkr.dto.accountDto.ResponseAccountList;
-import org.storkforge.barkr.mapper.AccountMapper;
-
-import java.util.List;
 
 @RestController
 @RequestMapping("/api")
